Add restore function to Global Translation
