+++ mmd_tools/operators/__init__.py
@@ -18,6 +18,7 @@ if "bpy" in locals():
     importlib.reload(rigid_body)
     importlib.reload(view)
     importlib.reload(sdef)
+    importlib.reload(translations)
 else:
     import bpy
     from . import (
@@ -33,5 +34,6 @@ else:
         morph,
         rigid_body,
         view,
-        sdef
+        sdef,
+        translations,
         )
