+++ mmd_tools/properties/root.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 """ MMDモデルパラメータ用Prop
 """
-from typing import Dict, List, Tuple
 import bpy
 from bpy.types import PropertyGroup
 from bpy.props import BoolProperty, CollectionProperty, IntProperty, StringProperty, EnumProperty
@@ -13,7 +12,7 @@ from mmd_tools.core.material import FnMaterial
 from mmd_tools.core.sdef import FnSDEF
 from mmd_tools.properties.morph import MaterialMorph, UVMorph, BoneMorph, VertexMorph, GroupMorph
 import mmd_tools.core.model as mmd_model
-from mmd_tools.translations import DictionaryEnum
+from mmd_tools.properties.translations import MMDDataQuery, MMDDataReference, MMDDataReferenceIndex
 
 def __driver_variables(id_data, path, index=-1):
     d = id_data.driver_add(path, index)
@@ -448,120 +447,3 @@ class MMDRoot(PropertyGroup):
         set=_setActiveMeshObject,
         get=_getActiveMeshObject,
         )
-
-MMD_DATA_TYPE_ENUM_ITEMS = [
-    (mmd_model.MMDDataType.BONE.name, mmd_model.MMDDataType.BONE.value, 'Bones', 1),
-    (mmd_model.MMDDataType.MORPH.name, mmd_model.MMDDataType.MORPH.value, 'Morphs', 2),
-    (mmd_model.MMDDataType.MATERIAL.name, mmd_model.MMDDataType.MATERIAL.value, 'Materials', 4),
-    (mmd_model.MMDDataType.DISPLAY.name, mmd_model.MMDDataType.DISPLAY.value, 'Display frames', 8),
-    (mmd_model.MMDDataType.PHYSICS.name, mmd_model.MMDDataType.PHYSICS.value, 'Rigidbodies and joints', 16),
-    (mmd_model.MMDDataType.INFO.name, mmd_model.MMDDataType.INFO.value, 'Model name and comments', 32),
-]
-
-@register_wrap
-class MMDDataReference(bpy.types.PropertyGroup):
-    type: bpy.props.EnumProperty(items=MMD_DATA_TYPE_ENUM_ITEMS)
-    object: bpy.props.PointerProperty(type=bpy.types.Object)
-    data_path: bpy.props.StringProperty()
-
-
-
-OPERATION_SCRIPT_PRESET_ITEMS: List[Tuple[str, str, str, int]] = [
-    ('NOTHING', '', '', 1),
-    ('TO_ENGLISH', 'Translate to English', 'to_english(name)', 2),
-    ('TO_MMD_LR', 'Blender L/R to MMD L/R', 'to_mmd_lr(name)', 3),
-    ('TO_BLENDER_LR', 'MMD L/R to Blender L/R', 'to_blender_lr(name_j)', 4),
-]
-
-@register_wrap
-class MMDDataQuery(bpy.types.PropertyGroup):
-    @staticmethod
-    def _update_index(mmd_data_query: 'MMDDataQuery', _context):
-        """Display the selected data in the Property Editor"""
-        if mmd_data_query.result_data_index < 0:
-            return
-
-        mmd_data_ref: MMDDataReference = mmd_data_query.result_data[mmd_data_query.result_data_index]
-
-        mmd_model.MMD_DATA_TYPE_TO_HANDLERS[mmd_data_ref.type].update_index(mmd_data_ref)
-
-    @staticmethod
-    def _update_query(mmd_data_query: 'MMDDataQuery', _context):
-        """Update the data by the query"""
-
-        mmd_data_query.result_data.clear()
-        mmd_data_query.result_data_index = -1
-
-        filter_japanese_blank: bool = mmd_data_query.filter_japanese_blank
-        filter_english_blank: bool = mmd_data_query.filter_english_blank
-
-        filter_selected: bool = mmd_data_query.filter_selected
-        filter_visible: bool = mmd_data_query.filter_visible
-
-        def check_data_visible(select: bool, hide: bool) -> bool:
-            return (
-                filter_selected and not select
-                or
-                filter_visible and hide
-            )
-
-        def check_blank_name(name_j: str, name_e: str) -> bool:
-            return (
-                filter_japanese_blank and name_j
-                or
-                filter_english_blank and name_e
-            )
-
-        for handler in mmd_model.MMD_DATA_HANDLERS:
-            if handler.type_name in mmd_data_query.filter_types:
-                handler.update_query(mmd_data_query, check_data_visible, check_blank_name)
-
-    @staticmethod
-    def _update_operation_script_preset(mmd_data_query: 'MMDDataQuery', _context):
-        if mmd_data_query.operation_script_preset == 'NOTHING':
-            return
-
-        id2scripts: Dict[str, str] = {i[0]:i[2] for i in OPERATION_SCRIPT_PRESET_ITEMS}
-
-        operation_script = id2scripts.get(mmd_data_query.operation_script_preset)
-        if operation_script is None:
-            return
-
-        mmd_data_query.operation_script = operation_script
-
-    filter_japanese_blank: bpy.props.BoolProperty(name='Japanese Blank', default=False, update=_update_query.__func__)
-    filter_english_blank: bpy.props.BoolProperty(name='English Blank', default=False, update=_update_query.__func__)
-    filter_selected: bpy.props.BoolProperty(name='Selected', default=False, update=_update_query.__func__)
-    filter_visible: bpy.props.BoolProperty(name='Visible', default=False, update=_update_query.__func__)
-    filter_types: bpy.props.EnumProperty(
-        items=MMD_DATA_TYPE_ENUM_ITEMS,
-        default={'BONE', 'MORPH', 'MATERIAL', 'DISPLAY', 'PHYSICS', },
-        options={'ENUM_FLAG'},
-        update=_update_query.__func__,
-    )
-    result_data_index: bpy.props.IntProperty(update=_update_index.__func__)
-    result_data: bpy.props.CollectionProperty(type=MMDDataReference)
-
-    dictionary = bpy.props.EnumProperty(
-        items=DictionaryEnum.get_dictionary_items,
-        name='Dictionary',
-    )
-
-    operation_target: bpy.props.EnumProperty(
-        items=[
-            ('BLENDER', 'Blender Name (name)', '', 1),
-            ('JAPANESE', 'Japanese Name (name_j)', '', 2),
-            ('ENGLISH', 'English Name (name_e)', '', 3),
-        ],
-        name='Operation Target',
-        default='JAPANESE',
-    )
-
-    operation_script_preset: bpy.props.EnumProperty(
-        items=OPERATION_SCRIPT_PRESET_ITEMS,
-        name='Operation Script Preset',
-        default='NOTHING',
-        update=_update_operation_script_preset.__func__,
-    )
-
-    operation_script: bpy.props.StringProperty()
